+++ platform/mv3/extension/js/mode-manager.js
@@ -223,7 +223,7 @@ export async function readFilteringModeDetails(bypassCache = false) {
         adminReadEx('noFiltering'),
     ]);
     if ( userModes === undefined ) {
-        userModes = { basic: [ 'all-urls' ] };
+        userModes = { optimal: [ 'all-urls' ] };
     }
     userModes = unserializeModeDetails(userModes);
     if ( Array.isArray(adminNoFiltering) ) {
@@ -347,9 +347,12 @@ export async function syncWithBrowserPermissions() {
     if ( beforeMode > MODE_BASIC && allowedHostnames.has('all-urls') === false ) {
         await setDefaultFilteringMode(MODE_BASIC);
         modified = true;
+    } else if ( beforeMode === MODE_BASIC && allowedHostnames.has('all-urls') ) {
+        await setDefaultFilteringMode(MODE_OPTIMAL);
+        modified = true;
     }
     const afterMode = await getDefaultFilteringMode();
-    if ( afterMode > MODE_BASIC ) { return false; }
+    if ( afterMode > MODE_BASIC ) { return afterMode !== beforeMode; }
     const filteringModes = await getFilteringModeDetails();
     const { optimal, complete } = filteringModes;
     for ( const hn of optimal ) {