+++ platform/mv3/extension/js/ruleset-manager.js
@@ -23,7 +23,6 @@ import {
     EXCLUDED_INITIATOR_DOMAINS,
     INITIATOR_DOMAINS,
     TAB_ID_NONE,
-    browser,
     dnr,
     i18n,
     localRead, localRemove, localWrite,
@@ -36,9 +35,9 @@ import {
     saveRulesetConfig,
 } from './config.js';
 
-
 import { fetchJSON } from './fetch.js';
 import { getAdminRulesets } from './admin.js';
+import { hasBroadHostPermissions } from './utils.js';
 import { ubolLog } from './debug.js';
 
 /******************************************************************************/
@@ -162,13 +161,7 @@ async function updateRemoveparamRules(currentRules, addRules, removeRuleIds) {
         removeRuleIds.push(rule.id);
     }
 
-    const [
-        hasOmnipotence,
-        rulesetDetails,
-    ] = await Promise.all([
-        browser.permissions.contains({ origins: [ '<all_urls>' ] }),
-        getEnabledRulesetsDetails(),
-    ]);
+    const rulesetDetails = await getEnabledRulesetsDetails();
 
     // Fetch removeparam rules for all enabled rulesets
     const toFetch = [];
@@ -180,12 +173,10 @@ async function updateRemoveparamRules(currentRules, addRules, removeRuleIds) {
 
     // Removeparam rules can only be enforced with omnipotence
     const allRules = [];
-    if ( hasOmnipotence ) {
-        for ( const rules of removeparamRulesets ) {
-            if ( Array.isArray(rules) === false ) { continue; }
-            for ( const rule of rules ) {
-                allRules.push(rule);
-            }
+    for ( const rules of removeparamRulesets ) {
+        if ( Array.isArray(rules) === false ) { continue; }
+        for ( const rule of rules ) {
+            allRules.push(rule);
         }
     }
     if ( allRules.length === 0 ) { return; }
@@ -207,13 +198,7 @@ async function updateRedirectRules(currentRules, addRules, removeRuleIds) {
         removeRuleIds.push(rule.id);
     }
 
-    const [
-        hasOmnipotence,
-        rulesetDetails,
-    ] = await Promise.all([
-        browser.permissions.contains({ origins: [ '<all_urls>' ] }),
-        getEnabledRulesetsDetails(),
-    ]);
+    const rulesetDetails = await getEnabledRulesetsDetails();
 
     // Fetch redirect rules for all enabled rulesets
     const toFetch = [];
@@ -225,12 +210,10 @@ async function updateRedirectRules(currentRules, addRules, removeRuleIds) {
 
     // Redirect rules can only be enforced with omnipotence
     const allRules = [];
-    if ( hasOmnipotence ) {
-        for ( const rules of redirectRulesets ) {
-            if ( Array.isArray(rules) === false ) { continue; }
-            for ( const rule of rules ) {
-                allRules.push(rule);
-            }
+    for ( const rules of redirectRulesets ) {
+        if ( Array.isArray(rules) === false ) { continue; }
+        for ( const rule of rules ) {
+            allRules.push(rule);
         }
     }
     if ( allRules.length === 0 ) { return; }
@@ -251,13 +234,7 @@ async function updateModifyHeadersRules(currentRules, addRules, removeRuleIds) {
         removeRuleIds.push(rule.id);
     }
 
-    const [
-        hasOmnipotence,
-        rulesetDetails,
-    ] = await Promise.all([
-        browser.permissions.contains({ origins: [ '<all_urls>' ] }),
-        getEnabledRulesetsDetails(),
-    ]);
+    const rulesetDetails = await getEnabledRulesetsDetails();
 
     // Fetch modifyHeaders rules for all enabled rulesets
     const toFetch = [];
@@ -269,12 +246,10 @@ async function updateModifyHeadersRules(currentRules, addRules, removeRuleIds) {
 
     // Redirect rules can only be enforced with omnipotence
     const allRules = [];
-    if ( hasOmnipotence ) {
-        for ( const rules of rulesets ) {
-            if ( Array.isArray(rules) === false ) { continue; }
-            for ( const rule of rules ) {
-                allRules.push(rule);
-            }
+    for ( const rules of rulesets ) {
+        if ( Array.isArray(rules) === false ) { continue; }
+        for ( const rule of rules ) {
+            allRules.push(rule);
         }
     }
     if ( allRules.length === 0 ) { return; }
@@ -350,7 +325,7 @@ async function updateStrictBlockRules(currentRules, addRules, removeRuleIds) {
         permanentlyExcluded = [],
         temporarilyExcluded = [],
     ] = await Promise.all([
-        browser.permissions.contains({ origins: [ '<all_urls>' ] }),
+        hasBroadHostPermissions(),
         getEnabledRulesetsDetails(),
         localRead('excludedStrictBlockHostnames'),
         sessionRead('excludedStrictBlockHostnames'),