+++ src/js/html-filtering.js
@@ -20,7 +20,7 @@
 */
 
 import { StaticExtFilteringHostnameDB } from './static-ext-filtering-db.js';
-import { entityFromDomain } from './uri-utils.js';
+import { entityFromHostname } from './uri-utils.js';
 import logger from './logger.js';
 import { sessionFirewall } from './filtering-engines.js';
 import µb from './background.js';
@@ -30,7 +30,7 @@ import µb from './background.js';
 const pselectors = new Map();
 const duplicates = new Set();
 
-const filterDB = new StaticExtFilteringHostnameDB(2);
+const filterDB = new StaticExtFilteringHostnameDB();
 
 let acceptedCount = 0;
 let discardedCount = 0;
@@ -260,7 +260,7 @@ function logOne(details, exception, selector) {
         .setDocOriginFromURL(details.url)
         .setFilter({
             source: 'extended',
-            raw: `${exception === 0 ? '##' : '#@#'}^${selector}`
+            raw: `${exception === 0 ? '##' : '#@#'}^${selector}`,
         })
         .toLogger();
 }
@@ -338,16 +338,11 @@ htmlFilteringEngine.compile = function(parser, writer) {
     let hasOnlyNegated = true;
     for ( const { hn, not, bad } of parser.getExtFilterDomainIterator() ) {
         if ( bad ) { continue; }
-        let kind = isException ? 0b01 : 0b00;
-        if ( not ) {
-            kind ^= 0b01;
-        } else {
+        const prefix = ((isException ? 1 : 0) ^ (not ? 1 : 0)) ? '-' : '+';
+        if ( not === false ) {
             hasOnlyNegated = false;
         }
-        if ( compiled.charCodeAt(0) === 0x7B /* '{' */ ) {
-            kind |= 0b10;
-        }
-        compiledFilters.push([ 64, hn, kind, compiled ]);
+        compiledFilters.push([ 64, hn, `${prefix}${compiled}` ]);
     }
 
     // Not allowed since it's equivalent to forbidden generic HTML filters
@@ -373,59 +368,45 @@ htmlFilteringEngine.fromCompiledContent = function(reader) {
         }
         duplicates.add(fingerprint);
         const args = reader.args();
-        filterDB.store(args[1], args[2], args[3]);
+        filterDB.store(args[1], args[2]);
     }
 };
 
 htmlFilteringEngine.retrieve = function(fctxt) {
-    const plains = new Set();
-    const procedurals = new Set();
-    const exceptions = new Set();
-    const retrieveSets = [ plains, exceptions, procedurals, exceptions ];
-
+    const all = new Set();
     const hostname = fctxt.getHostname();
-    filterDB.retrieve(hostname, retrieveSets);
-
-    const domain = fctxt.getDomain();
-    const entity = entityFromDomain(domain);
-    const hostnameEntity = entity !== ''
-        ? `${hostname.slice(0, -domain.length)}${entity}`
-        : '*';
-    filterDB.retrieve(hostnameEntity, retrieveSets, 1);
-
-    if ( plains.size === 0 && procedurals.size === 0 ) { return; }
+    filterDB.retrieveSpecifics(all, hostname);
+    const entity = entityFromHostname(hostname, fctxt.getDomain());
+    filterDB.retrieveSpecifics(all, entity);
+    filterDB.retrieveSpecificsByRegex(all, hostname, fctxt.url);
+    filterDB.retrieveGenerics(all);
+    if ( all.size === 0 ) { return; }
 
     // https://github.com/gorhill/uBlock/issues/2835
     //   Do not filter if the site is under an `allow` rule.
-    if (
-        µb.userSettings.advancedUserEnabled &&
-        sessionFirewall.evaluateCellZY(hostname, hostname, '*') === 2
-    ) {
-        return;
+    if ( µb.userSettings.advancedUserEnabled ) {
+        if ( sessionFirewall.evaluateCellZY(hostname, hostname, '*') === 2 ) { return; }
     }
 
-    const out = { plains, procedurals };
-
-    if ( exceptions.size === 0 ) {
-        return out;
-    }
-
-    for ( const selector of exceptions ) {
-        if ( plains.has(selector) ) {
-            plains.delete(selector);
-            logOne(fctxt, 1, selector);
-            continue;
-        }
-        if ( procedurals.has(selector) ) {
-            procedurals.delete(selector);
-            logOne(fctxt, 1, JSON.parse(selector).raw);
-            continue;
+    // Split filters in different groups
+    const plains = new Set();
+    const procedurals = new Set();
+    for ( const s of all ) {
+        if ( s.charCodeAt(0) === 0x2D /* - */ ) { continue; }
+        const selector = s.slice(1);
+        const isProcedural = selector.startsWith('{');
+        if ( all.has(`-${selector}`) ) {
+            logOne(fctxt, 1, isProcedural ? JSON.parse(selector).raw : selector);
+        } else if ( isProcedural ) {
+            procedurals.add(selector);
+        } else {
+            plains.add(selector);
         }
     }
 
-    if ( plains.size !== 0 || procedurals.size !== 0 ) {
-        return out;
-    }
+    if ( plains.size === 0 && procedurals.size === 0 ) { return; }
+
+    return { plains, procedurals };
 };
 
 htmlFilteringEngine.apply = function(doc, details, selectors) {