+++ src/js/scriptlet-filtering.js
@@ -37,10 +37,8 @@ import µb from './background.js';
 
 /******************************************************************************/
 
-const contentScriptRegisterer = new (class {
-    constructor() {
-        this.hostnameToDetails = new Map();
-    }
+const contentScriptRegisterer = {
+    hostnameToDetails: new Map(),
     register(hostname, code) {
         if ( browser.contentScripts === undefined ) { return false; }
         if ( hostname === '' ) { return false; }
@@ -66,7 +64,7 @@ const contentScriptRegisterer = new (class {
         });
         this.hostnameToDetails.set(hostname, { handle: promise, code });
         return false;
-    }
+    },
     unregister(hostname) {
         if ( hostname === '' ) { return; }
         if ( this.hostnameToDetails.size === 0 ) { return; }
@@ -74,7 +72,7 @@ const contentScriptRegisterer = new (class {
         if ( details === undefined ) { return; }
         this.hostnameToDetails.delete(hostname);
         this.unregisterHandle(details.handle);
-    }
+    },
     flush(hostname) {
         if ( hostname === '' ) { return; }
         if ( hostname === '*' ) { return this.reset(); }
@@ -84,14 +82,14 @@ const contentScriptRegisterer = new (class {
             if ( pos !== 0 && hn.charCodeAt(pos-1) !== 0x2E /* . */ ) { continue; }
             this.unregister(hn);
         }
-    }
+    },
     reset() {
         if ( this.hostnameToDetails.size === 0 ) { return; }
         for ( const details of this.hostnameToDetails.values() ) {
             this.unregisterHandle(details.handle);
         }
         this.hostnameToDetails.clear();
-    }
+    },
     unregisterHandle(handle) {
         if ( handle instanceof Promise ) {
             handle.then(handle => {
@@ -100,8 +98,8 @@ const contentScriptRegisterer = new (class {
         } else {
             handle.unregister();
         }
-    }
-})();
+    },
+};
 
 /******************************************************************************/
 
@@ -370,17 +368,16 @@ export class ScriptletFilteringEngineEx extends ScriptletFilteringEngine {
     toLogger(request, details) {
         if ( details === undefined ) { return; }
         if ( logger.enabled !== true ) { return; }
-        if ( typeof details.filters !== 'string' ) { return; }
-        const fctxt = µb.filteringContext
+        if ( Array.isArray(details.filters) === false ) { return; }
+        µb.filteringContext
             .duplicate()
             .fromTabId(request.tabId)
             .setRealm('extended')
             .setType('scriptlet')
             .setURL(request.url)
-            .setDocOriginFromURL(request.url);
-        for ( const raw of details.filters.split('\n') ) {
-            fctxt.setFilter({ source: 'extended', raw }).toLogger();
-        }
+            .setDocOriginFromURL(request.url)
+            .setFilter(details.filters.map(a => ({ source: 'extended', raw: a })))
+            .toLogger();
     }
 }